fix: when login is removed
